+++ src/headers/read-agents.h
@@ -21,6 +21,7 @@ typedef struct _agent_info {
     char *rootcheck_endtime;
     char *os;
     char *version;
+    char *merged_sum;
 } agent_info;
 
 /* Print syscheck db (of modified files) */